Updated openmpi formula for wider generic support.
Configure flags are now modified on a case-by-case basis depending on version
name and target architechure.

Versions passed to the smithy formula install invocation can be suffixed with
'-mtm' to trigger the build to include mpi-thread-multiple configure options
without need to modify the formula.

If the architecture is detected to be 'dtn', then the '--with-verb' options are
not invoked as the DTNs should not be expected to have infiniband connections.
This should be extended to any arch that does not use infiniband.